Update drain_queues to check if queue has already been drained
When GeventScheduler is being used as async_.schedule there is a chance
the queues are being drained while a drain_queue job is scheduled to be
executed on the gevent event loop. This is valid so change assert is_tick_used
to check if queue has already been drained instead of throwing an assertion
exception.