minor grammer changes
Thanks for you work! I was reading the docs and noticed a few minor things. one of them looked like this "`obj` act" I wasn't sure if act is a term I am not familiar with or if it was intended to be the end of object so I took a guess :man_shrugging:.